implement vote Api test an Class test
